Makefile updates

+++ system/gdt.c
@@ -58,7 +58,7 @@ void gdt_install()
 { 
 	/* Setup the GDT pointer */
 	gp.limit = (sizeof(struct gdt_entry) * 3) - 1; 
-	gp.base = &gdt; 
+	gp.base = (int)&gdt; 
 
 	/* Null descriptor */
 	gdt_set_gate(0, 0, 0, 0, 0);
